requirements.txt
aiohttp
apscheduler
bs4
cryptg
ffmpeg-python
gallery-dl
google-generativeai
hachoir
hydrogram
pixivpy3
Pillow
pydantic
psycopg2-binary
pysaucenao
python-dotenv
pytesseract
PyYaml
PyPDF2
pytz
rich
sqlalchemy
streamlit==1.39.0
telegraph
telegraph[aio]
telethon
tgcrypto
typer
yt-dlp